TB(API)/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
PyJWT
aiomysql
websockets
aiohttp
google-auth
requests
httpx